# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube Channel Feed API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Channel {
  id: number;
  youtubeChannelId: string;
  name: string;
  thumbnailUrl?: string | null;
  addedAt: string;
}

export interface AddChannelRequest {
  /** YouTube channel ID (starts with UC...) or handle (@username) */
  youtubeChannelId: string;
}

export interface Video {
  videoId: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  publishedAt: string;
  viewCount?: string | null;
  channelId: string;
  channelName: string;
  channelThumbnailUrl?: string | null;
  duration?: string | null;
}

export interface ChannelSearchResult {
  youtubeChannelId: string;
  name: string;
  description: string;
  thumbnailUrl?: string | null;
  subscriberCount?: string | null;
}

export interface VideoSummaryRequest {
  videoId: string;
  title: string;
  description: string;
  channelName: string;
}

export interface VideoSummaryResponse {
  summary: string;
}

export interface ErrorResponse {
  error: string;
}

export type SearchChannelsParams = {
  /**
   * Search query
   */
  q: string;
};

export type ListVideosParams = {
  /**
   * Filter by specific channel DB id
   */
  channelId?: number;
};

export type SearchBeatChannelsParams = {
  q: string;
};

export type ListBeatsParams = {
  channelId?: number;
};

export type GetSimilarBeatsParams = {
  title: string;
};