# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YouTube Channel Feed API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: channels
    description: Favourite YouTube channels
  - name: videos
    description: YouTube videos from favourite channels
  - name: ai
    description: AI-powered features
  - name: beats
    description: Beat channels and beat videos
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List favourite channels
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: addChannel
      tags: [channels]
      summary: Add a favourite channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddChannelRequest"
      responses:
        "201":
          description: Channel added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /channels/search:
    get:
      operationId: searchChannels
      tags: [channels]
      summary: Search YouTube channels by name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelSearchResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /channels/{id}:
    delete:
      operationId: removeChannel
      tags: [channels]
      summary: Remove a favourite channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Channel removed
        "404":
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List recent videos from favourite channels
      parameters:
        - name: channelId
          in: query
          required: false
          schema:
            type: integer
          description: Filter by specific channel DB id
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
        "500":
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /beat-channels:
    get:
      operationId: listBeatChannels
      tags: [beats]
      summary: List beat channels
      responses:
        "200":
          description: List of beat channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: addBeatChannel
      tags: [beats]
      summary: Add a beat channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddChannelRequest"
      responses:
        "201":
          description: Beat channel added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /beat-channels/search:
    get:
      operationId: searchBeatChannels
      tags: [beats]
      summary: Search YouTube channels for beat producers
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelSearchResult"
  /beat-channels/{id}:
    delete:
      operationId: removeBeatChannel
      tags: [beats]
      summary: Remove a beat channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Beat channel removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /beats:
    get:
      operationId: listBeats
      tags: [beats]
      summary: List recent beats from beat channels
      parameters:
        - name: channelId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of beats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
        "500":
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /beats/{videoId}/similar:
    get:
      operationId: getSimilarBeats
      tags: [beats]
      summary: Get similar beats for a given video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
        - name: title
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Similar beats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
  /videos/summary:
    post:
      operationId: getVideoSummary
      tags: [ai]
      summary: Generate AI summary for a video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoSummaryRequest"
      responses:
        "200":
          description: AI-generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoSummaryResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Channel:
      type: object
      properties:
        id:
          type: integer
        youtubeChannelId:
          type: string
        name:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - youtubeChannelId
        - name
        - addedAt
    AddChannelRequest:
      type: object
      properties:
        youtubeChannelId:
          type: string
          description: YouTube channel ID (starts with UC...) or handle (@username)
      required:
        - youtubeChannelId
    Video:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        publishedAt:
          type: string
          format: date-time
        viewCount:
          type: string
          nullable: true
        channelId:
          type: string
        channelName:
          type: string
        channelThumbnailUrl:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
      required:
        - videoId
        - title
        - description
        - thumbnailUrl
        - publishedAt
        - channelId
        - channelName
    ChannelSearchResult:
      type: object
      properties:
        youtubeChannelId:
          type: string
        name:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        subscriberCount:
          type: string
          nullable: true
      required:
        - youtubeChannelId
        - name
        - description
    VideoSummaryRequest:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        description:
          type: string
        channelName:
          type: string
      required:
        - videoId
        - title
        - description
        - channelName
    VideoSummaryResponse:
      type: object
      properties:
        summary:
          type: string
      required:
        - summary
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error